+++ src/process/process_pcdrs.py
@@ -39,12 +39,12 @@ class ProcessPcdrs(ProcessBase):
             "offset": {
                 "data": np.empty(self.shapes["offset"]),
                 "path": "offset",
-                "type": np.int16
+                "type": np.int
             },
             "gainlevel_mean": {
                 "data": np.empty(self.shapes["offset"]),
                 "path": "gainlevel_mean",
-                "type": np.int16
+                "type": np.int
             },
             "slope": {
                 "data": np.empty(self.shapes["offset"]),
